+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+## [2.191.0](https://github.com/kiva/ui/compare/v2.190.2...v2.191.0) (2022-06-16)
+
+
+### Features
+
+* zero loans reset analytics ([47f4c2c](https://github.com/kiva/ui/commit/47f4c2c0cf38c545fa90254b36499f2f17bf0706))
+
+### [2.190.2](https://github.com/kiva/ui/compare/v2.190.1...v2.190.2) (2022-06-16)
+
+
+### Bug Fixes
+
+* selected amount was added to view tracking for complete loan test ([b0a5bc8](https://github.com/kiva/ui/commit/b0a5bc8e7a1c7f47c01c77d2484b5c60b0c79463))
+
+### [2.190.1](https://github.com/kiva/ui/compare/v2.190.0...v2.190.1) (2022-06-16)
+
+
+### Bug Fixes
+
+* update layout of mobile vs desktop filter interface ([dc69963](https://github.com/kiva/ui/commit/dc6996300a7f3b74f22718ec9f3038dab3b74b07))
+
+## [2.190.0](https://github.com/kiva/ui/compare/v2.189.0...v2.190.0) (2022-06-16)
+
+
+### Features
+
+* added 0 loans state ([95b30f9](https://github.com/kiva/ui/commit/95b30f944ddb65d675fcdf3fc41e796cb5098fad))
+* added results per page analytics ([2db1d97](https://github.com/kiva/ui/commit/2db1d9790712098bea2591931f68f8c37650013c))
+* results per page component ([77b4d5f](https://github.com/kiva/ui/commit/77b4d5fcc48e50b947ec9140e4bc206897c80857))
+
 ## [2.189.0](https://github.com/kiva/ui/compare/v2.188.2...v2.189.0) (2022-06-16)